Refactor: - implement integration test for account repository
