+++ assignment-bank-api/src/test/java/com/rbank/bank/controller/AccountControllerIntegrationTest.java
@@ -1,0 +1,24 @@
+package com.rbank.bank.controller;
+
+import com.rbank.bank.dao.AccountRepository;
+import com.rbank.bank.model.Account;
+import com.rbank.bank.model.User;
+import com.rbank.util.BaseIntegrationTest;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.any;
+
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+class AccountControllerIntegrationTest extends BaseIntegrationTest {
+
+
+
+
+
+
+}
