+++ assignment-bank-api/src/test/java/com/rbank/util/BaseUnitTest.java
@@ -1,0 +1,20 @@
+package com.rbank.util;
+
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.testcontainers.containers.PostgreSQLContainer;
+
+@ActiveProfiles("test")
+@Tag("unit")
+@ExtendWith(SpringExtension.class)
+public class BaseUnitTest {
+
+
+}
